+++ packages/admin/admin-theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/admin-theme
 
+## 7.4.2
+
+### Patch Changes
+
+-   @comet/admin-icons@7.4.2
+
+## 7.4.1
+
+### Patch Changes
+
+-   @comet/admin-icons@7.4.1
+
 ## 7.4.0
 
 ### Patch Changes