+++ packages/api/cms-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/cms-api
 
+## 7.4.2
+
+### Patch Changes
+
+-   @comet/blocks-api@7.4.2
+
+## 7.4.1
+
+### Patch Changes
+
+-   @comet/blocks-api@7.4.1
+
 ## 7.4.0
 
 ### Minor Changes