+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/eslint-config
 
+## 7.4.2
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.4.2
+
+## 7.4.1
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.4.1
+
 ## 7.4.0
 
 ### Patch Changes